feat: add mapstruct and exception handlers

+++ src/main/java/org/example/proxibanque/service/impl/ClientServiceImpl.java
@@ -3,6 +3,12 @@ package org.example.proxibanque.service.impl;
 import org.example.proxibanque.dto.request.ClientCreateRequest;
 import org.example.proxibanque.dto.request.CreateRunningAccountRequest;
 import org.example.proxibanque.dto.request.TransferRequest;
+import org.example.proxibanque.dto.response.ClientResponse;
+import org.example.proxibanque.exception.custom.ClientNotFoundException;
+import org.example.proxibanque.exception.custom.InvalidDataException;
+import org.example.proxibanque.exception.custom.NoAccountException;
+import org.example.proxibanque.exception.custom.NotEnoughMoneyException;
+import org.example.proxibanque.mapper.ClientMapper;
 import org.example.proxibanque.model.entity.Agency;
 import org.example.proxibanque.model.entity.Client;
 import org.example.proxibanque.model.entity.RunningAccount;
@@ -18,92 +24,95 @@ import java.util.Optional;
 public class ClientServiceImpl implements ClientService {
 
     private final AgencyServiceHelper agencyServiceHelper;
-    private ClientRepository clientRepository;
+    private final ClientRepository clientRepository;
+    private final ClientMapper clientMapper;
 
-    public ClientServiceImpl(ClientRepository clientRepository, AgencyServiceHelper agencyServiceHelper) {
+    public ClientServiceImpl(ClientRepository clientRepository, AgencyServiceHelper agencyServiceHelper, ClientMapper clientMapper) {
         this.clientRepository = clientRepository;
         this.agencyServiceHelper = agencyServiceHelper;
+        this.clientMapper = clientMapper;
     }
 
     @Override
-    public Client createClient(ClientCreateRequest request) {
-        Optional<Agency> agencyOpt = agencyServiceHelper.getAgency(request.getAgencyId());
-        // Un systeme de globalHandler pour les exceptions seraient necessaires mais pas le temps
+    public ClientResponse createClient(ClientCreateRequest request) {
+        Optional<Agency> agencyOpt = agencyServiceHelper.getAgency(request.agencyId());
         if (agencyOpt.isEmpty()) {
-            return null;
+            throw new InvalidDataException();
         }
         Agency agency = agencyOpt.get();
-        Client client = new Client(request.getNom(), request.getPrenom(), request.getAdresse(),  request.getCodePostal(), request.getVille(), request.getTelephone(), agency);
+        Client client = new Client(request.nom(), request.prenom(), request.adresse(),  request.codePostal(), request.ville(), request.telephone(), agency);
         clientRepository.save(client);
-        return client;
+        return clientMapper.clientToDto(client);
     }
 
     @Override
-    public List<Client> getAllClients() {
-        return  clientRepository.findAll();
+    public List<ClientResponse> getAllClients() {
+        return  clientRepository.findAll().stream().map(clientMapper::clientToDto).toList();
     }
 
     @Override
     public void deleteClient(Long id) {
+        if (clientRepository.existsById(id)) {
+            throw new ClientNotFoundException();
+        }
         clientRepository.deleteById(id);
     }
 
     @Override
-    public Optional<Client> getClientById(Long id) {
-        return clientRepository.findById(id);
+    public ClientResponse getClientById(Long id) {
+        return clientRepository.findById(id).map(clientMapper::clientToDto).orElseThrow(ClientNotFoundException::new);
     }
 
     @Override
-    public Client openAccount(Long clientId, CreateRunningAccountRequest accountDto) {
+    public ClientResponse openAccount(Long clientId, CreateRunningAccountRequest accountDto) {
         Client client = clientRepository.getClientById(clientId);
-        RunningAccount runningAccount = new RunningAccount(accountDto.getInitialAmount());
+        RunningAccount runningAccount = new RunningAccount(accountDto.initialAmount());
         client.setRunningAccount(runningAccount);
         clientRepository.save(client);
-        return client;
+        return clientMapper.clientToDto(client);
     }
 
     @Override
-    public boolean transferMoney(Long clientId, Long targetId, TransferRequest transferRequest) {
+    public void transferMoney(Long clientId, Long targetId, TransferRequest transferRequest) {
         Client client = clientRepository.getClientById(clientId);
         Client targetClient = clientRepository.getClientById(targetId);
         if (client == null || targetClient == null)
-            return false;
+            throw new InvalidDataException();
         if (client.getRunningAccount() ==  null || targetClient.getRunningAccount() == null) {
-            return false;
+            throw new InvalidDataException();
         }
-        if (!client.getRunningAccount().withdraw(transferRequest.getAmount())) {
-            return false;
+        if (!client.getRunningAccount().withdraw(transferRequest.amount())) {
+            throw new NotEnoughMoneyException();
         }
-        targetClient.getRunningAccount().deposit(transferRequest.getAmount());
+        targetClient.getRunningAccount().deposit(transferRequest.amount());
         clientRepository.save(targetClient);
         clientRepository.save(client);
-        return true;
     }
 
     @Override
-    public Client deposit(Long clientId, TransferRequest transferRequest) {
+    public ClientResponse deposit(Long clientId, TransferRequest transferRequest) {
         Client client = clientRepository.getClientById(clientId);
         if  (client == null)
-            return null;
+            throw new InvalidDataException();
         RunningAccount runningAccount = client.getRunningAccount();
         if (runningAccount == null)
-            return null;
-        runningAccount.deposit(transferRequest.getAmount());
+            throw new NoAccountException();
+        runningAccount.deposit(transferRequest.amount());
         clientRepository.save(client);
-        return client;
+        return clientMapper.clientToDto(client);
     }
 
     @Override
-    public Client withdraw(Long clientId, TransferRequest transferRequest) {
+    public ClientResponse withdraw(Long clientId, TransferRequest transferRequest) {
         Client client = clientRepository.getClientById(clientId);
         if (client == null)
-            return null;
+            throw new InvalidDataException();
         RunningAccount runningAccount = client.getRunningAccount();
         if (runningAccount == null)
-            return null;
-        if (!runningAccount.withdraw(transferRequest.getAmount()))
-            return null;
+            throw new NoAccountException();
+        if (!runningAccount.withdraw(transferRequest.amount()))
+            throw new NotEnoughMoneyException();
         clientRepository.save(client);
-        return client;
+        return clientMapper.clientToDto(client);
     }
 }
